ai-services/planner/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
sqlalchemy
pymysql
cryptography
